Update assig5_8.c
Yes. Your question is right. However, we are writing programs to understand and apply, basic concepts of an array, base address, pointer, loop, function etc. Dynamic arrays are possible in C as well as C++. In C programmer has to manage it using dynamic memory allocation. The concept has been dealt in a separate chapter ahead in the curriculum. 

Further, as far as this program is concerned, declare two arrays and write the program since array size is small.